Refactored parser to keep archiver as a separate concern
Added 'deploy' to top level makefile
