+++ parser/.coveragerc
@@ -1,3 +1,4 @@
 [report]
 exclude_lines =
     if __name__ == .__main__.:
+    if 'AWS_REGION' in os.environ:
