+++ parser/parser/__init__.py
@@ -1,0 +1,2 @@
+""" Parse and archive inbound person data
+"""
